create Viterbi class, rewrite KoreanTokenizer
